gerando url presign para upload arquivo

+++ src/main/java/org/erick/file_flow/dto/URLResponse.java
@@ -4,6 +4,6 @@ public record URLResponse(
         Long jobId, 
         Long documentId, 
         String uploadURL, 
-        Integer expiresInSeconds
+        Long expiresInSeconds
     ) {
 }
